requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
requests
openai

